init repositories

+++ src/main/java/com/miguelsouza/libraryapi/model/Livro.java
@@ -1,8 +1,10 @@
 package com.miguelsouza.libraryapi.model;
 
+import com.miguelsouza.libraryapi.model.enums.GeneroLivro;
 import jakarta.persistence.*;
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.util.UUID;
 
@@ -30,7 +32,7 @@ public class Livro {
     private GeneroLivro genero;
 
     @Column(name = "preco", precision = 18, scale = 2)
-    private Double preco;
+    private BigDecimal preco;
 
     @ManyToOne
     @JoinColumn(name = "id_autor")
